refactor: extract upload core CSS to styles folder
