+++ packages/upload/src/styles/vaadin-upload-file-styles.d.ts
