+++ packages/upload/src/vaadin-upload-file.js
@@ -9,8 +9,8 @@ import { html, LitElement, nothing } from 'lit';
 import { defineCustomElement } from '@vaadin/component-base/src/define.js';
 import { PolylitMixin } from '@vaadin/component-base/src/polylit-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import { uploadFileStyles } from './styles/vaadin-upload-file-styles.js';
 import { UploadFileMixin } from './vaadin-upload-file-mixin.js';
-import { uploadFileStyles } from './vaadin-upload-file-styles.js';
 
 /**
  * `<vaadin-upload-file>` element represents a file in the file list of `<vaadin-upload>`.
