+++ packages/upload/src/vaadin-upload.js
@@ -7,12 +7,13 @@ import '@vaadin/button/src/vaadin-button.js';
 import './vaadin-upload-icon.js';
 import './vaadin-upload-icons.js';
 import './vaadin-upload-file-list.js';
-import { css, html, LitElement } from 'lit';
+import { html, LitElement } from 'lit';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { defineCustomElement } from '@vaadin/component-base/src/define.js';
 import { ElementMixin } from '@vaadin/component-base/src/element-mixin.js';
 import { PolylitMixin } from '@vaadin/component-base/src/polylit-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import { uploadStyles } from './styles/vaadin-upload-styles.js';
 import { UploadMixin } from './vaadin-upload-mixin.js';
 
 /**
@@ -70,21 +71,7 @@ class Upload extends UploadMixin(ElementMixin(ThemableMixin(PolylitMixin(LitElem
   }
 
   static get styles() {
-    return css`
-      :host {
-        display: block;
-        position: relative;
-        box-sizing: border-box;
-      }
-
-      :host([hidden]) {
-        display: none !important;
-      }
-
-      [hidden] {
-        display: none !important;
-      }
-    `;
+    return uploadStyles;
   }
 
   /** @protected */
